legion: One more fix for 64-bit points

+++ runtime/realm/indexspace.h
@@ -303,7 +303,7 @@ namespace Realm {
         for (int i = 0; i < MAX_POINT_DIM; i++)
           point_data[i] = 0;
       }
-      DomainPoint(int index) : dim(0)
+      DomainPoint(off_t index) : dim(0)
       {
         point_data[0] = index;
         for (int i = 1; i < MAX_POINT_DIM; i++)
